# pb/fxgateway.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "fxgateway.proto",
    "version": "version not set"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/system/function/{FunctionName}": {
      "get": {
        "operationId": "GetMeta",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/pbFunction"
            }
          }
        },
        "parameters": [
          {
            "name": "FunctionName",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "FxGateway"
        ]
      },
      "delete": {
        "operationId": "Delete",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/pbMessage"
            }
          }
        },
        "parameters": [
          {
            "name": "FunctionName",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "FxGateway"
        ]
      }
    },
...
    "/system/scale-function": {
      "put": {
        "operationId": "ReplicaUpdate",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/pbMessage"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbScaleServiceRequest"
            }
          }
        ],
        "tags": [
          "FxGateway"
        ]
      }
    }
  },
  "definitions": {
...
    "pbFunction": {
      "type": "object",
      "properties": {
        "Name": {
          "type": "string"
        },
        "Image": {
          "type": "string"
        },
        "InvocationCount": {
          "type": "string",
          "format": "uint64"
        },
        "Replicas": {
          "type": "string",
          "format": "uint64"
        },
        "AvailableReplicas": {
          "type": "string",
          "format": "uint64"
        },
        "Annotations": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        },
        "Labels": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        }
      }
    },
...
    "pbMessage": {
      "type": "object",
      "properties": {
        "Msg": {
          "type": "string"
        }
      }
    },
...
    "pbScaleServiceRequest": {
      "type": "object",
      "properties": {
        "ServiceName": {
          "type": "string"
        },
        "Replicas": {
          "type": "string",
          "format": "uint64"
        }
      }
    }
  }
}